+++ CHANGELOG.md
@@ -1,14 +1,19 @@
-## 0.1.2
- -  Fix duplicate requests when getting a language
+## 0.1.3
+ - Rename open api
+ - Fix URI duplicate link error
 
 ## 0.1.2
- -  Fix error package name in doc
+ - Fix duplicate requests when getting a language
+
+## 0.1.2
+ - Fix error package name in doc
 
 ## 0.1.1
 
 - Rename class method
 - Add fetch languages method to class and example,test
 - Rename package name
+
 ## 0.1.0
 
 - Initial version, created by Stagehand